fixed derivation in equation 14 and flow on
fixed derivation in equation 14, acknowledging the speculative manner of trialling the theta gradient based scaling factor and fixed all eqns and captions with d/ds to reflect the actual expressions in eqn 16 & 18. The graphs do not change as the correct expressions were used in the calculations, it was the paper that needed fixing.